pyproject.toml
[project]
name = "sinau"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.2",
    "django-cors-headers>=4.7.0",
    "django-debug-toolbar>=5.2.0",
    "django-extensions>=4.1",
    "django-import-export[all]>=4.3.7",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-material>=0.3.3",
    "drf-restwind>=0.2.4",
    "drf-standardized-errors>=0.14.1",
    "environs>=14.2.0",
    "eskiz-sms>=0.2.3",
    "gunicorn>=23.0.0",
    "icecream>=2.1.4",
    "phonenumbers>=9.0.6",
    "pillow>=11.2.1",
    "pytest>=8.4.0",
    "pytest-factoryboy>=2.7.0",
    "python-decouple>=3.8",
    "twilio>=9.6.2",
    "uvicorn>=0.34.3",
    "whitenoise>=6.9.0",
]
